chore: refactor auth methods
